requirements.txt
-i https://pypi.org/simple
certifi==2023.5.7; python_version >= '3.6'
charset-normalizer==3.2.0; python_full_version >= '3.7.0'
idna==3.4; python_version >= '3.5'
pytz==2023.3
pyyaml==6.0
requests==2.31.0
urllib3==2.0.3; python_version >= '3.7'
